v0.0.12
- Bug fix.
- Added a training section.

+++ Data/Scripts/Global/Inventory/Armor.cs
@@ -6,11 +6,11 @@ public partial class Armor : Item
 {
     [Export] public string ArmorType { get; set; }
     [Export] public float Protection { get; set; }
-    [Export] public float AdditionalHealth { get; set; }
+    [Export] public int AdditionalHealth { get; set; }
 
     public Armor() : base() { }
 
-    public Armor(int id, int maxCount, string itemName, string description, float protection, float additionalHealth, string armorType) : base(id, maxCount, itemName, description)
+    public Armor(int id, int maxCount, string itemName, string description, float protection, int additionalHealth, string armorType) : base(id, maxCount, itemName, description)
     {
         Protection = protection;
         AdditionalHealth = additionalHealth;
